requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
binance==0.3
certifi==2023.5.7
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
numpy==1.25.1
pandas==2.0.3
psycopg2-binary==2.9.6
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
telebot==0.0.5
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2
